requirements.txt
click==8.0.3
colorama==0.4.4
DateTime==4.3
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.10.0
itsdangerous==2.0.1
Jinja2==3.0.3
linecache2==1.0.0
MarkupSafe==2.0.1
PyJWT==1.7.1
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.29
traceback2==1.4.0
typing_extensions==4.0.1
unittest2==1.1.0
uuid==1.30
Werkzeug==2.0.2
zipp==3.7.0
zope.interface==5.4.0
